requirements.txt
pandas==2.1.4
numpy==1.26.3
pyyaml==6.0.1
python-dateutil==2.8.2
